Fixes typos, grammar, etc in interface fidelity post.
I really like this post. :) The idea of reducing the interface dilemma to the problem of the production of quantization artifacts is a neat thought. The fact that interfaces can restrict expression is the key point that needs to be hit upon again and again and again. Good stuff! Will link back!